refactor: use typescript
