+++ .env.example
@@ -1,0 +1,2 @@
+PORT=
+WEBSOCKET_PORT=
