+++ package.json
@@ -2,20 +2,35 @@
     "name": "ws_task",
     "version": "1.0.0",
     "description": "Template for the WebSocket remote control task",
-    "main": "index.js",
-    "type": "module",
+    "main": "src/index.ts",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "start:dev": "nodemon src/index.ts",
+        "start": "npm run clean:dist && npm run build && node ./dist/index.js",
+        "clean:dist": "rm -rf ./dist",
+        "build": "npx tsc --outDir 'dist'",
+        "lint": "eslint . --ext .ts --ignore-path .gitignore --max-warnings 0 --fix"
     },
     "author": "Andrei Auchynnikau",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/ws": "^8.5.3",
+        "@typescript-eslint/eslint-plugin": "^5.29.0",
+        "@typescript-eslint/parser": "^5.29.0",
+        "eslint": "^8.18.0",
+        "eslint-config-airbnb-base": "^15.0.0",
+        "eslint-config-airbnb-typescript": "^17.0.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-eslint-comments": "^3.2.0",
+        "eslint-plugin-import": "^2.26.0",
+        "eslint-plugin-prettier": "^4.0.0",
+        "nodemon": "^2.0.18",
+        "prettier": "^2.7.1",
+        "ts-node": "^10.8.1",
+        "typescript": "^4.7.4"
     },
     "dependencies": {
+        "dotenv": "^16.0.1",
         "jimp": "^0.16.1",
         "robotjs": "^0.6.0",
         "ws": "^8.8.0"
